[ver 1.01.86] Refactoring TktkContainer Code

+++ TktkDirectX12GameLib/TktkContainerLib/inc/TktkContainer/ResourceContainer/ResourceContainerInitParam.h
@@ -6,8 +6,8 @@ namespace tktkContainer
 	// 「ResourceContainer」のインスタンスを作る時に必要な変数
 	struct ResourceContainerInitParam
 	{
-		unsigned int staticNodeNum			{ 10U };	// 初期化時にインスタンス何個分のメモリを事前確保するか？
-		unsigned int moveNodeMaxNumPerFrame	{ 1U };		// １フレームで初期化時確保メモリに移動できる要素数
+		size_t	staticNodeNum			{ 10U };	// 初期化時にインスタンス何個分のメモリを事前確保するか？
+		size_t	moveNodeMaxNumPerFrame	{ 1U };		// １フレームで初期化時確保メモリに移動できる要素数
 	};
 }
 #endif // !RESOURCE_CONTAINER_INIT_PARAM_H_
